Major changes to hardware, added software. Current version locks
Added MATLAB and Python software copied from a previous dispersive
feedback project.  Modified said files so that they work with the
new project.

Lots of testing of VHDL with testbenches and waveform files.
There is now a topmod_tb file for testing the overall design.

Design will lock the phase in the envisaged setup with the DAC
outputs going through multipliers and then the laser fields going
through double-pass AOMs.  This generates a frequency signal that
is 8 times the difference frequency programmed into the device.
In order to lock, I had to divide the actuator phase from the
control component PhaseControl.vhd by 2^8 rather than the 2^4 I
expected.  I suspect that I added an extra division by 2 somewhere
in the VHDL code.