Major changes to bit widths for phase calculations
Changed the DDS output widths and set the DDS compilers to use
a Taylor series correction to try and fix sawtooth phase measure-
ments.  This necessitated an increase in the bit widths for the CIC
decimators and the multipliers.  While it seemed to reduce the
amount of noise in the phase measurement, locking is less stable
in that it is likely to unlock.  Additionally, I now see weird
phase jumps when the system is locked.  I do not understand what
the problem might be.