Many minor changes to VHDL code.  Using new software classes
Can't remember exactly what has changed.  This version uses two
clocks in the design rather than the single clock in the previous
version with the clock-convertor in the AXI logic.

Been having lots of trouble with both measuring the phase and
implementing control.  Might be an issue with the clock domains.

PI controller needs to be changed as it is not quite right: the
division by 2^N occurs BEFORE the accumulator for the phase, and
it has to occur after to avoid truncating the ability of the PI
controller to correct for errors.  I think I will change this
module to reflect my new understanding about pipelined servo
modules.  I think I will also change it so that the output of the
module is always of the same type as the input measurement; that
is to say, it will be a signed phase from -pi to pi rather than
a signed phase from 0 to 2*pi.

I have added a function "convertPhase" which does a simpler job
of converting between the signed and unsigned phase representations.
It's based on the observation that the conversion from signed to
unsigned integers, and vice versa, follows exactly the same pattern
as the conversion of a phase in the domain -pi to pi to 0 to 2*pi.