Added debugging features, fixed major error in phase unwrapping
Fixed error in phase unwrapping in PhaseControl.vhd.  The constant
representing pi had not been changed since I changed the phase
calculation width from 16 bits to 24 bits.  This meant that if the
phase jumped from -pi to pi (or probably vice versa) it didn't
properly account for it.  This constant has been changed, and it
seems as though this has made the phase locking more robust and
less prone to developing spontaneous oscillation on abrupt phase
changes.

In order to test this with a physical device, I have added a few
features that allow me to produce a single frequency on both outputs
and demodulate at that frequency, even when the timing controller
is generating the outputs.

I have changed the PI controller to a PID controller and it now
operates as a properly pipelined device.  Not that this is needed
since the phase calculation itself takes many cycles, but it is
still useful.

I have also added an output multiplier for amplitude control.  This
has a 12 bit input value to multiply the 10 bit DDS output.
Amplitude control has been added to the timing controller.

Finally, I created a little GUI for easy visualization of the
phase lock parameters.