Added hold ability to servo through timing controller flags
Users can now upload flags to timing controller FIFO so that the
PID controller can be enabled/disabled and the current DDS output
can be held during a timing sequence.  There are 4 bits to the flag
signal, with flag(0) being the enable flag and flag(1) being the
hold flag.  The enable flag is used to ensure that the servo
locks to an unwrapped phase equal to 0.  When the hold flag is
'1' the PID controller stops accumulating the actuator value
so that the output value is held constant.  This also prevents
integral windup of the PID output.

Test results are mixed when applying an interferometer sequence
with 3 Gaussian pulses of FWHM of 30 us 1 ms apart.  When the
pulses are not applied the output is set to a 5 or 25 MHz signal
with an amplitude of 0.1, and the phase lock is engaged.  Sometimes
the lock re-engages at an unwrapped phase of 0 (as it should) and
sometimes it locks elsewhere (not sure how that's possible).
Sometimes that lock oscillates.  It might be another phase boundary
crossing issue.