Fixed issue with servo hold that prevented proper re-locking
When the servo (PhaseControl.vhd) is in its "hold" mode the phase
unwrapping process no longer accumulates the phase difference. This
means that when the calculated phase does all sorts of weird shit
because there is insufficient amplitude to properly calculate the
phase the unwrapped phase does not wander off to the wop-wops.
This means that the unwrapped phase, which is the input to the
PID controller, remains near its old value which means the lock
is more easily re-acquired.

I have tested this with 30 us FWHM Gaussian pulses, with a static
frequency of 0.5 MHz, and holding over +/- 3*width and the lock
re-acquires nicely and I can change the phase between pulses.