Moved design over to SIGNALlab 250-12 device, still problems
Moved phase lock design over to the SIGNALlab 250-12 device, which
has ADCs and DACs that operate at 250 MSPS.  Mostly have things
ported over, but still having some issues with the phase lock when
using the timing controller.  Probably just the timing controller
acting up.

Need to switch the timing controller over from the FIFO to a RAM,
so that the same information is used everytime and we don't have
to upload on every run.  Maybe this will fix the issues?

Also, the level of noise on the phase measurement with the SIGNALlab
seems higher than for the STEMlab 125-14 device.  Not sure if that's
just inherent to the device or if it's poor design on my part.

Added a block memory to save ADC data on this device for debugging.