Added external triggering (falling edge), added calibration files
Devices can now be externally triggered on the falling edge of an
input signal.  Had to add an extra parameter to turn off the ext.
triggering ability so that the current software trigger still works.

Also added some calibration .mat files for converting normalized
optical powers to amplitudes, commensurate with the multipliers +
amplifiers in use.

Added a second kind of Bragg function where the lock is enabled
only for 500 us before the pulses arrive.  This reduces spontaneous
emission problems and atom loss during the "dark" phases of the
interferometer.  Currently set to turn on 500 us before the pulses,
and the lock seems to acquire just fine during this time.