Added many comments to rtl files, started README.
I added a bunch of comments to the rtl files because I haven't worked
on this project in a long time, so I wanted to remind myself how it
worked.  I also **started** a README, but I haven't finished it yet.
Mostly I can't remember the commands for compiling programs on the
SIGNALlab device, so I don't want to write something wrong.