Started to merge the STEMlab and SIGNALlab designs
The designs for both boards are pretty similar and it doesn't
make any sense to keep them as separate projects.  Instead,
I am going to try and merge them together with different folders
for files for different boards.  Directories with suffixes -125 are
for the STEMlab board and with suffixes -250 are for the SIGNALlab
board.